services/ops-api/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
psycopg2-binary>=2.9.9,<3.0.0
pydantic>=2.8.0,<3.0.0
itsdangerous>=2.2.0,<3.0.0
python-multipart>=0.0.20,<1.0.0
cryptography>=45.0.0,<46.0.0
boto3>=1.39.0,<2.0.0
Pillow>=11.0.0,<12.0.0